# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:55556","description":"Generated server url"}],"paths":{"/reserva":{"post":{"tags":["reserva-controller"],"operationId":"createReserva","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/requestPostDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reserva"}}}}}}},"/reserva/{id_reserva}":{"get":{"tags":["reserva-controller"],"operationId":"getReservaById","parameters":[{"name":"id_reserva","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reserva"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"requestPostDTO":{"type":"object","properties":{"id_hotel":{"type":"integer","format":"int32"},"id_cliente":{"type":"string"},"tipo_habitacion":{"type":"string"},"fecha_ingreso":{"type":"string","format":"date"},"fecha_salida":{"type":"string","format":"date"},"medio_pago":{"type":"string"}}},"Reserva":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"id_hotel":{"type":"integer","format":"int32"},"id_cliente":{"type":"string"},"tipo_habitacion":{"type":"string"},"fecha_ingreso":{"type":"string","format":"date"},"fecha_salida":{"type":"string","format":"date"},"medio_pago":{"type":"string"},"precio":{"type":"number","format":"double"},"estado_reserva":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}